Moved transpileLabelStmt and transpileGotoStmt into goto.go
