+++ transpiler/transpiler.go
@@ -351,30 +351,3 @@ func transpileStmts(nodes []ast.Node, p *program.Program) ([]goast.Stmt, error) 
 
 	return stmts, nil
 }
-
-func transpileLabelStmt(n *ast.LabelStmt, p *program.Program) (*goast.LabeledStmt, error) {
-	var stmt goast.Stmt
-	if len(n.Children()) > 0 {
-		var err error
-		stmt, _, _, err = transpileToStmt(n.Children()[0], p)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if stmt == nil {
-		stmt = &goast.EmptyStmt{}
-	}
-
-	return &goast.LabeledStmt{
-		Label: util.NewIdent(n.Name),
-		Stmt:  stmt,
-	}, nil
-}
-
-func transpileGotoStmt(n *ast.GotoStmt, p *program.Program) (*goast.BranchStmt, error) {
-	return &goast.BranchStmt{
-		Label: util.NewIdent(n.Name),
-		Tok:   token.GOTO,
-	}, nil
-}
